# cmd/api/docs/swagger.yaml
definitions:
  domain.Category:
    properties:
      category:
        type: string
      id:
        type: integer
    type: object
  domain.Product:
    properties:
      category:
        $ref: '#/definitions/domain.Category'
      category_id:
        type: integer
      color:
        type: string
      id:
        type: integer
      price:
        type: number
      product_name:
        type: string
      stock:
        type: integer
      url:
        type: string
    type: object
...
  models.ProductUpdate:
    properties:
      product_id:
        type: integer
      stock:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Show Times is an E-commerce platform to purchase Watch
  title: Go + Gin E-Commerce API Show Times
  version: 1.0.0
paths:
...
  /admin/product:
    delete:
      consumes:
      - application/json
      description: Deletes an existing product using its ID.
      parameters:
      - description: ID of the product to be deleted
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Invalid product ID or deletion error'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Delete product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Fetches a paginated list of products.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of products per page (default: 5)'
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product list retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Error displaying products'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: List products
      tags:
      - Product
    patch:
      consumes:
      - application/json
      description: Updates an existing product using the provided details.
      parameters:
      - description: Inventory ID of the product to update
        in: query
        name: inventory_id
        required: true
        type: integer
      - description: Updated product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Invalid input or product update error'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Edit product
      tags:
      - Product
    post:
      consumes:
      - multipart/form-data
      description: Adds a new product using the provided details and image.
      parameters:
      - description: Category ID of the product
        in: formData
        name: category_id
        required: true
        type: integer
      - description: Name of the product
        in: formData
        name: product_name
        required: true
        type: string
      - description: Color of the product
        in: formData
        name: color
        required: true
        type: string
      - description: Stock quantity of the product
        in: formData
        name: stock
        required: true
        type: integer
      - description: Price of the product
        in: formData
        name: price
        required: true
        type: number
      - description: Image file of the product
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Product added successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Invalid input or image upload error'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Add a new product
      tags:
      - Product
...
  /product/stock:
    patch:
      consumes:
      - application/json
      description: Updates the stock quantity of an existing product.
      parameters:
      - description: Updated stock details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ProductUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Product stock updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Invalid input or update error'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Update product stock
      tags:
      - Product
...
securityDefinitions:
  BearerTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"